requirements.txt
python-telegram-bot
pillow
google-generativeai
python-dotenv
schedule
requests
telebot
pytelegrambotapi
psycopg2-binary==2.9.9 
